Create SortingList
I was running several lines of code and this managed to work. I had a question though:
If I were to replace line two by writing "lst_b = lst_a", the code does not function properly.
Why does this happen? Another way I can ask this question is why I have to establish two lists with the same terms, or if there is a way to do it without this method.